Revert "fix committee test"
This reverts commit 7a0931c4487707e3eaf75a592415577f04253772.

+++ beacon-chain/cache/committee_test.go
@@ -78,13 +78,13 @@ func TestCommitteeCache_ActiveCount(t *testing.T) {
 	item := &Committees{Seed: [32]byte{'A'}, SortedIndices: []uint64{1, 2, 3, 4, 5, 6}}
 	count, err := cache.ActiveIndicesCount(item.Seed)
 	require.NoError(t, err)
-	assert.Equal(t, uint64(0), count, "Expected active count not to exist in empty cache")
+	assert.Equal(t, 0, count, "Expected active count not to exist in empty cache")
 
 	require.NoError(t, cache.AddCommitteeShuffledList(item))
 
 	count, err = cache.ActiveIndicesCount(item.Seed)
 	require.NoError(t, err)
-	assert.Equal(t, uint64(len(item.SortedIndices)), count)
+	assert.Equal(t, len(item.SortedIndices), count)
 }
 
 func TestCommitteeCache_CanRotate(t *testing.T) {
